Implement AHP Ranking Algorithm and Handlers
- Added AlgorithmAhpHandler to handle the AHP ranking logic.
- Created IndexRankingKepalaSekolahHandler for managing rankings based on school head evaluations.
- Introduced ShowRankingHandler to display rankings for specific evaluations.
- Developed TotalBobotKriteriaHandler to calculate total weights for criteria.
- Established RankingService to provide a unified interface for ranking operations.
- Implemented necessary model relationships and data retrieval for ranking calculations.
